Add more unit tests for header buttons
