+++ packages/wp-story-editor/src/components/statusCheck/test/statusCheck.js
@@ -23,11 +23,11 @@ import {
 } from '@testing-library/react';
 import { setAppElement } from '@web-stories-wp/design-system';
 import { ConfigContext } from '@web-stories-wp/story-editor';
-import { renderWithTheme } from '@web-stories-wp/story-editor/src/testUtils';
 
 /**
  * Internal dependencies
  */
+import { renderWithTheme } from '../../../testUtils';
 import StatusCheck from '../statusCheck';
 
 jest.mock('../../../api/statusCheck');
